2024.03.28   FIX bulding issues in GCC

+++ src/viewers/events/traingapstatdialog.cpp
@@ -48,7 +48,7 @@ void TrainGapStatModel::setupModel()
         it->setData(p->second.begin()->operator*().secs(),Qt::EditRole);
         setItem(row,ColValue,it);
         it=new SI;
-        it->setData(p->second.size(),Qt::EditRole);
+        it->setData(static_cast<int>(p->second.size()),Qt::EditRole);
         setItem(row,ColCount,it);
         row++;
     }
